Code for comparison with KamLAND-Zen
Main script KLZComp.py added, which allows you to set limits on all Majoron
modelsfor KamLAND-Zen (KLZ). The script also reproduces the KLZ spectral plot
either with the original KLZ limits or the 90% CLs calculated.

In order to calculate the 90% CLs the set_limit module had to be updated, first
to be compatible with the new Isotope class and second to re-design the way
limits on half life were calculated. I now calculate the 90% CL on the null
hypothesis (i.e. just background), by plotting delta-chi-squared as a function
of inverse half life.

Small changes to the isotope, reconstruced and physics modules were required
through developing the code. Alsp the constants and defaults modules were
update with the relevant information for Xe136 and KLZ analysis.

Finally added a new module plots, to handle macros for generating the standard
plots that MajoRat can make.