Adding function to plot majoron spectrum
Function added to plots.py to format and produce spectral plot for Majorons.
Also includes an additional function in error_utils to check if two values are
equal to a given precision. This is employed in isotope.py to ensure the sum of
the masses of each scintillator component equals the mass of scintillator.